feat(fw): add auto calibration routine to pca9685 driver
With this features, one can measure the real frequency of the internal
oscillator of the pca9685 in order to have more precise control.
This procedure requires one of the channels to be connected to a gpio
pin which will then be used to measure the real frequency.

Once the routine has finished, the real value is applied and is also
printed.

This is based from
https://github.com/adafruit/Adafruit-PWM-Servo-Driver-Library/blob/master/examples/oscillator/oscillator.ino